+++ cmd/create/appcr/command.go
@@ -5,10 +5,12 @@ import (
 )
 
 var flag struct {
+	Annotations         []string
 	AppName             string
 	AppNamespace        string
 	AppVersion          string
 	Catalog             string
+	ConfigMajorVersion  int
 	DisableForceUpgrade bool
 	Name                string
 	Output              string
@@ -23,12 +25,14 @@ func NewCommand() *cobra.Command {
 		RunE:  runAppCRError,
 	}
 
+	cmd.Flags().StringSliceVar(&flag.Annotations, "annotations", nil, "list of annotations to add to the generated App CR in `name=value,...` format")
 	cmd.Flags().StringVar(&flag.AppName, "app-name", "", "app name")
 	cmd.Flags().StringVar(&flag.AppNamespace, "app-namespace", "giantswarm", "app namespace")
 	cmd.Flags().StringVar(&flag.AppVersion, "app-version", "", "app version")
 	cmd.Flags().StringVar(&flag.Catalog, "catalog", "", "app catalog name")
+	cmd.Flags().IntVar(&flag.ConfigMajorVersion, "config-major-version", 0, "major version of giantswarm/config to use for this App CR")
 	cmd.Flags().BoolVar(&flag.DisableForceUpgrade, "disable-force-upgrade", false, "disable helm chart force upgrade")
-	cmd.Flags().StringVar(&flag.Name, "name", "", "cr name")
+	cmd.Flags().StringVar(&flag.Name, "name", "", "CR name")
 	cmd.Flags().StringVarP(&flag.Output, "output", "o", "yaml", "output format. allowed: yaml,json")
 	cmd.Flags().StringVar(&flag.UserConfigMapName, "user-configmap-name", "", "user configmap name")
 	cmd.Flags().StringVar(&flag.UserSecretName, "user-secret-name", "", "user secret name")